requirements.txt
python-telegram-bot>=20.2
python-dotenv>=0.15.0
validators>=0.20.0
yt-dlp>=2023.3.4
